[FIX] payment_ogone: ogone expect the amount to be an integer
Oversight of python3 compat changes made at 2e6a589f417af48e3322955ffff6e94b6d8a24e3

+++ addons/payment_ogone/models/payment.py
@@ -344,7 +344,7 @@ class PaymentTxOgone(models.Model):
             'USERID': account.ogone_userid,
             'PSWD': account.ogone_password,
             'ORDERID': reference,
-            'AMOUNT': self.amount * 100,
+            'AMOUNT': int(self.amount * 100),
             'CURRENCY': self.currency_id.name,
             'OPERATION': 'SAL',
             'ECI': 2,   # Recurring (from MOTO)
